+++ source/App/source/ExampleHost.FunctionApp02/Program.cs
@@ -17,7 +17,7 @@ using Energinet.DataHub.Core.App.FunctionApp.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 
 var host = new HostBuilder()
-    .ConfigureFunctionsWorkerDefaults()
+    .ConfigureFunctionsWebApplication()
     .ConfigureServices(services =>
     {
         // Configuration verified in tests. See comments in FunctionApp01.Program.