+++ source/App/source/ExampleHost.WebApi01/Startup.cs
@@ -40,7 +40,7 @@ public class Startup
         //  * We can see Trace, Request, Dependencies and other entries in App Insights out-of-box.
         //    See https://docs.microsoft.com/en-us/azure/azure-monitor/app/asp-net-core
         //  * Telemetry events are enriched with property "Subsystem" and configured value
-        services.AddApplicationInsightsForWebApp(subsystemName: "ExampleHost.WebApp");
+        services.AddApplicationInsightsForWebApp(subsystemName: "ExampleHost.WebApi");
 
         // Configure HttpClient for calling WebApi02
         services.AddHttpClient(HttpClientNames.WebApi02, httpClient =>
@@ -54,7 +54,7 @@ public class Startup
 
         // Swagger and api versioning (verified in tests)
         services
-            .AddSwaggerForWebApp(Assembly.GetExecutingAssembly(), swaggerUITitle: "ExampleHost.WebApp")
+            .AddSwaggerForWebApp(Assembly.GetExecutingAssembly(), swaggerUITitle: "ExampleHost.WebApi")
 
             // Setting default version to 2.0, this will be overwritten if the method has it's own version
             .AddApiVersioningForWebApp(new ApiVersion(2, 0));