+++ source/App/documents/release-notes/release-notes.md
@@ -1,5 +1,15 @@
 # App Release notes
 
+## Version 13.0.0
+
+- See also [Version 13.0.0 release notes](./version_13_0_0.md)
+- In `FunctionApp` project:
+    - Refactored implementation to only work with [ASP.NET Core integration for HTTP](https://learn.microsoft.com/en-us/azure/azure-functions/dotnet-isolated-process-guide?tabs=windows#aspnet-core-integration).
+    - Added functionality to configure authentication and authorization for HttpTrigger's
+- In `WebApp` project:
+    - Removed the obsolete overload of `AuthenticationExtensions.AddJwtBearerAuthenticationForWebApp`
+- Moved type `Energinet.DataHub.Core.App.WebApp.Extensions.Options.UserAuthenticationOptions` to `Energinet.DataHub.Core.App.Common.Extensions.Options.UserAuthenticationOptions`.
+
 ## Version 12.2.1
 
 - In `FunctionApp` project: